add: dto

+++ src/main/java/org/simple/sm/quartz/service/impl/JobServiceImpl.java
@@ -1,19 +1,22 @@
 package org.simple.sm.quartz.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import org.simple.sm.common.base.BaseResultDTO;
+import org.simple.sm.common.constant.ConstantColumn;
 import org.simple.sm.common.enumeration.ENUM_BASE_RESULT;
 import org.simple.sm.common.enumeration.ENUM_JOB_STATUS;
 import org.simple.sm.component.snowflake.IdWorker;
 import org.simple.sm.db.sqlite.entity.TQuartzInfo;
 import org.simple.sm.db.sqlite.service.TQuartzInfoService;
+import org.simple.sm.quartz.dto.req.GetJobsReqDTO;
 import org.simple.sm.quartz.dto.req.JobReqDTO;
-import org.simple.sm.quartz.dto.res.JobResDTO;
+import org.simple.sm.quartz.dto.req.UpdateJobReqDTO;
+import org.simple.sm.quartz.dto.res.GetJobsResDTO;
 import org.simple.sm.quartz.service.JobService;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
 import javax.annotation.Resource;
-import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
@@ -26,20 +29,22 @@ public class JobServiceImpl implements JobService {
     TQuartzInfoService tQuartzInfoService;
 
     @Override
-    public JobResDTO addJob(JobReqDTO jobReqDTO) {
-        JobResDTO jobResDTO = new JobResDTO();
+    public BaseResultDTO<?> addJob(JobReqDTO jobReqDTO) {
+        BaseResultDTO<?> resultDTO = new BaseResultDTO<>();
         jobReqDTO.setJobNo(String.valueOf(idWorker.nextId()));
         if (StringUtils.isEmpty(jobReqDTO.getJobName())
                 || StringUtils.isEmpty(jobReqDTO.getJobGroup())
-                || StringUtils.isEmpty(jobReqDTO.getExpression())) {
-            jobResDTO.failure(ENUM_BASE_RESULT.FAIL);
-            return jobResDTO;
+                || StringUtils.isEmpty(jobReqDTO.getExpression())
+                || StringUtils.isEmpty(jobReqDTO.getJobType())) {
+            resultDTO.failure(ENUM_BASE_RESULT.FAIL);
+            return resultDTO;
         }
         //db
         TQuartzInfo tQuartzInfo = new TQuartzInfo();
         tQuartzInfo.setJobNo(String.valueOf(idWorker.nextId()));
         tQuartzInfo.setJobGroup(jobReqDTO.getJobGroup());
         tQuartzInfo.setJobName(jobReqDTO.getJobName());
+        tQuartzInfo.setJobType(jobReqDTO.getJobType());
         tQuartzInfo.setExpression(jobReqDTO.getExpression());
         tQuartzInfo.setStatus(ENUM_JOB_STATUS.STOP.getCode());
         tQuartzInfo.setGmtCreate(new Date());
@@ -47,44 +52,54 @@ public class JobServiceImpl implements JobService {
         tQuartzInfo.setIsDelete(0);
         tQuartzInfoService.getBaseMapper().insert(tQuartzInfo);
         // todo add to local cache
-        jobResDTO.success(ENUM_BASE_RESULT.SUCCESS);
-        return jobResDTO;
+        resultDTO.success();
+        return resultDTO;
     }
     @Override
-    public List<JobResDTO> getJobs(JobReqDTO jobReqDTO) {
-        List<JobResDTO> jobResDTOs = new ArrayList<>();
+    public BaseResultDTO<GetJobsResDTO> getJobs(GetJobsReqDTO getJobsReqDTO) {
+        BaseResultDTO<GetJobsResDTO> resultDTO = new BaseResultDTO<>();
+        GetJobsResDTO getJobsResDTO = new GetJobsResDTO();
         QueryWrapper<TQuartzInfo> wrapper = new QueryWrapper<>();
         List<TQuartzInfo> tQuartzInfos = tQuartzInfoService.getBaseMapper().selectList(wrapper);
-        tQuartzInfos.forEach(e->{
-                    JobResDTO jobResDTO = new JobResDTO();
-                    jobResDTO.setJobNo(e.getJobNo());
-                    jobResDTO.setJobNo(e.getJobName());
-                    jobResDTO.setJobNo(e.getJobGroup());
-                    jobResDTO.setJobNo(e.getExpression());
-                    jobResDTOs.add(jobResDTO);
-                });
-        return jobResDTOs;
+        getJobsResDTO.setQuartzInfoList(tQuartzInfos);
+        resultDTO.success(getJobsResDTO);
+        return resultDTO;
     }
     @Override
-    public void updateJob(JobReqDTO jobReqDTO) {
+    public BaseResultDTO<?> updateJob(UpdateJobReqDTO updateJobReqDTO) {
+        BaseResultDTO<?> resultDTO = new BaseResultDTO<>();
+        if (StringUtils.isEmpty(updateJobReqDTO.getJobNo())) {
+            resultDTO.failure(ENUM_BASE_RESULT.PARAM_INPUT_ERR);
+            return resultDTO;
+        }
         //db
         TQuartzInfo tQuartzInfo = new TQuartzInfo();
-        tQuartzInfo.setJobName(jobReqDTO.getJobName());
-        tQuartzInfo.setJobGroup(jobReqDTO.getJobGroup());
-        tQuartzInfo.setStatus(jobReqDTO.getStatus());
-        tQuartzInfo.setExpression(jobReqDTO.getExpression());
+        tQuartzInfo.setJobName(updateJobReqDTO.getJobName());
+        tQuartzInfo.setJobGroup(updateJobReqDTO.getJobGroup());
+        tQuartzInfo.setJobType(updateJobReqDTO.getJobType());
+        tQuartzInfo.setStatus(updateJobReqDTO.getStatus());
+        tQuartzInfo.setExpression(updateJobReqDTO.getExpression());
         tQuartzInfo.setGmtModified(new Date());
         QueryWrapper<TQuartzInfo> wrapper = new QueryWrapper<>();
-        wrapper.eq("job_no",jobReqDTO.getJobNo());
+        wrapper.eq("job_no",updateJobReqDTO.getJobNo());
         tQuartzInfoService.getBaseMapper().update(tQuartzInfo,wrapper);
         // todo sync local cache
+        resultDTO.success();
+        return resultDTO;
     }
 
     @Override
-    public void removeJob(JobReqDTO jobReqDTO) {
+    public  BaseResultDTO<?> removeJob(JobReqDTO jobReqDTO) {
+        BaseResultDTO<?> resultDTO = new BaseResultDTO<>();
+        if (StringUtils.isEmpty(jobReqDTO.getJobNo())) {
+            resultDTO.failure(ENUM_BASE_RESULT.PARAM_INPUT_ERR);
+            return resultDTO;
+        }
         QueryWrapper<TQuartzInfo> wrapper = new QueryWrapper<>();
-        wrapper.eq("job_no",jobReqDTO.getJobNo());
+        wrapper.eq(ConstantColumn.JOB_NO,jobReqDTO.getJobNo());
         tQuartzInfoService.getBaseMapper().delete(wrapper);
         // todo sync local cache
+        resultDTO.success();
+        return resultDTO;
     }
 }
